+++ src/main/java/com/api/pizza/repository/IIssueRepository.java
@@ -1,6 +1,9 @@
 package com.api.pizza.repository;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +12,7 @@ import com.api.pizza.entity.Department;
 import com.api.pizza.entity.Equipment;
 import com.api.pizza.entity.Issue;
 
-public interface IIssueRepository extends JpaRepository<Issue, Integer> {
+public interface IIssueRepository extends JpaRepository<Issue, Long> {
 
     List<Issue> findAll(Specification<Issue> specification);
 
@@ -22,6 +25,5 @@ public interface IIssueRepository extends JpaRepository<Issue, Integer> {
     long countIssuesByDepartment(Department department);
 
     List<Issue> findByEquipmentId(Long equipmentIds);
-
 
 }
